Eliminate warnings on vc++ compiler (#2243)
Eliminate lots of Windows compiler warnings.

+++ filters/ChipperFilter.hpp
@@ -61,7 +61,7 @@ class PDAL_DLL ChipPtRef
 private:
     double m_pos;
     point_count_t m_ptindex;
-    uint32_t m_oindex;
+    size_t m_oindex;
 
 public:
     bool operator < (const ChipPtRef& pt) const
@@ -102,7 +102,7 @@ private:
     {
         return m_vec.end();
     }
-    ChipPtRef& operator[](uint32_t pos)
+    ChipPtRef& operator[](size_t pos)
     {
         return m_vec[pos];
     }
